requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.0
python-socketio==5.10.0
ebooklib==0.18
beautifulsoup4==4.12.2
soundfile==0.12.1
torch>=2.0.0
kokoro-onnx
Werkzeug==3.0.0
spacy>=3.7.0
num2words>=0.5.12
pysbd>=0.3.4
deep-translator>=1.11.4